Refactor TypeCheck a lot
Now typechecking cases that introduce a new context or lookup in the current context are separated from the cases that preserve the current context. The next step is to pass enough information through the functor that things like `ensureTerm` become trivial and no more access to the context is needed.