Overhaul error type and applicative
Now the recursion models typechecking directly: it will give access to the terms at the next level of recursion, as well as its type, and the type of its type, etc. (iff those computations succeed, of course).

In addition, the error type is more precise, and will capture the cursor without using `Mu` and `VariantF`. Well, it will, once I implement that. (Currently disabled until I develop the theory of recursive `Expr` zippers.)