Fix alphanorm, typecheck ctx, and tc app
Alpha normalization was broken for unknown reasons, so I converted it to a more direct algorithm: keep track of the simplest context, pass it down, just like the other algorithms. Might need to fix an edge case there ...

Typechecking fn applications was broken due to a silly mistake: substituting the type instead of the value.

Keeping track of variables in context forgot to shift in the new expression(s) being introed.