Share work during normalization
It's more shared _and_ more lazy! Now the context is normalized on demand. The change alone appears to have shaved 20s off of the run time of the tests. (120 -> 100) The rest of it didn't seem to make a big difference on its own, but of course it was the framework to allow that to happen.