New partial monoidal state monad-like thing
Code is a mess right now but the basic idea is that we accumulate
errors and good state in the same monoid, but split it out for
the purpose of knowing when an error has occurred, etc.

This will allow fancy things while unifying universes, rows,
you name it. Plus it means that state flows bottom-up.

Definitely lots of improvements to be made but this is a start.